+++ packages/semantic-release-clean-package-json/src/index.ts
@@ -1,21 +1,37 @@
-import { writeJson } from "@visulima/fs";
-import { join } from "@visulima/path";
+import { rm } from "node:fs/promises";
+
+import { isAccessible, readJson, writeJson } from "@visulima/fs";
+import { join, resolve } from "@visulima/path";
+import type { PackageJson } from "type-fest";
 
 import defaultKeepProperties from "./default-keep-properties";
-import type { PrepareContext } from "./definitions/context";
+import type { CommonContext, PublishContext } from "./definitions/context";
 import type { PluginConfig } from "./definitions/plugin-config";
 import getPackage from "./utils/get-pkg";
 
-// eslint-disable-next-line sonarjs/cognitive-complexity,import/prefer-default-export
-export const prepare = async (pluginConfig: PluginConfig, context: PrepareContext): Promise<void> => {
+// eslint-disable-next-line sonarjs/cognitive-complexity
+export const publish = async (pluginConfig: PluginConfig, context: PublishContext): Promise<void> => {
     const packageJson = await getPackage(pluginConfig, context);
+    const cwd = pluginConfig.pkgRoot ? resolve(context.cwd, pluginConfig.pkgRoot) : context.cwd;
+
+    await writeJson(
+        join(cwd, "package.json.back"),
+        packageJson,
+        {
+            detectIndent: true,
+        },
+    );
+
+    context.logger.log("Created a backup of the package.json file.");
 
     const keepProperties = new Set([...defaultKeepProperties, ...(pluginConfig.keep ?? [])]);
 
     context.logger.log(`Keeping the following properties: ${[...keepProperties].join(", ")}`);
 
+    const packageJsonCopy = { ...packageJson };
+
     // eslint-disable-next-line no-loops/no-loops,no-restricted-syntax
-    for (const property in packageJson) {
+    for (const property in packageJsonCopy) {
         if (keepProperties.has(property)) {
             // eslint-disable-next-line no-continue
             continue;
@@ -23,17 +39,17 @@ export const prepare = async (pluginConfig: PluginConfig, context: PrepareContex
 
         if (property === "scripts") {
             // eslint-disable-next-line no-loops/no-loops,no-restricted-syntax
-            for (const script in packageJson.scripts) {
+            for (const script in packageJsonCopy.scripts) {
                 if (keepProperties.has(`${property}.${script}`)) {
                     // eslint-disable-next-line no-continue
                     continue;
                 }
 
                 // eslint-disable-next-line @typescript-eslint/no-dynamic-delete,security/detect-object-injection
-                delete packageJson.scripts[script];
+                delete packageJsonCopy.scripts[script];
             }
 
-            if (packageJson.scripts && Object.keys(packageJson.scripts).length > 0) {
+            if (packageJsonCopy.scripts && Object.keys(packageJsonCopy.scripts).length > 0) {
                 // eslint-disable-next-line no-continue
                 continue;
             }
@@ -41,10 +57,40 @@ export const prepare = async (pluginConfig: PluginConfig, context: PrepareContex
 
         context.logger.log(`Removing property "${property}"`);
         // eslint-disable-next-line @typescript-eslint/no-dynamic-delete,security/detect-object-injection
-        delete packageJson[property];
+        delete packageJsonCopy[property];
     }
 
-    await writeJson(join(context.cwd, "package.json"), packageJson, {
+    await writeJson(join(cwd, "package.json"), packageJsonCopy, {
         detectIndent: true,
     });
+};
+
+export const success = async (pluginConfig: PluginConfig, context: CommonContext): Promise<void> => {
+    const cwd = pluginConfig.pkgRoot ? resolve(context.cwd, pluginConfig.pkgRoot) : context.cwd;
+
+    const backupPackageJson = join(cwd, "package.json.back");
+
+    if (await isAccessible(backupPackageJson)) {
+        const packageJson = await getPackage(pluginConfig, context);
+
+        const backupPackageJsonContent = (await readJson(backupPackageJson)) as PackageJson;
+
+        // Overwrite the version from the backup package.json
+        backupPackageJsonContent.version = packageJson.version;
+
+        await writeJson(
+            join(cwd, "package.json"),
+            backupPackageJsonContent,
+            {
+                detectIndent: true,
+                overwrite: true,
+            },
+        );
+
+        await rm(backupPackageJson);
+
+        context.logger.log("Restored modified package.json from backup.");
+    } else {
+        context.logger.error("No backup package.json found.");
+    }
 };